dev: mv mt_dev.c to dev dir
