+++ lib/src/datapath/mt_shared_queue.c
@@ -4,7 +4,7 @@
 
 #include "mt_shared_queue.h"
 
-#include "../mt_dev.h"
+#include "../dev/mt_dev.h"
 #include "../mt_flow.h"
 #include "../mt_log.h"
 #include "../mt_stat.h"
