+++ lib/src/dev/meson.build
@@ -1,0 +1,6 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright 2023 Intel Corporation
+
+sources += files(
+  'mt_dev.c',
+)
