+++ lib/src/dev/mt_dev.h
@@ -5,7 +5,7 @@
 #ifndef _MT_LIB_DEV_HEAD_H_
 #define _MT_LIB_DEV_HEAD_H_
 
-#include "mt_main.h"
+#include "../mt_main.h"
 
 /* default desc nb for tx and rx */
 #define MT_DEV_RX_DESC (4096 / 2)
@@ -32,9 +32,6 @@ int mt_dev_free(struct mtl_main_impl* impl);
 
 int mt_dev_start(struct mtl_main_impl* impl);
 int mt_dev_stop(struct mtl_main_impl* impl);
-
-int mt_dev_dst_ip_mac(struct mtl_main_impl* impl, uint8_t dip[MTL_IP_ADDR_LEN],
-                      struct rte_ether_addr* ea, enum mtl_port port, int timeout_ms);
 
 struct mt_tx_queue* mt_dev_get_tx_queue(struct mtl_main_impl* impl, enum mtl_port port,
                                         struct mt_txq_flow* flow);
@@ -73,7 +70,6 @@ static inline uint16_t mt_dpdk_rx_burst(struct mt_rx_queue* queue,
 
 int mt_dev_if_init(struct mtl_main_impl* impl);
 int mt_dev_if_uinit(struct mtl_main_impl* impl);
-int mt_dev_if_post_init(struct mtl_main_impl* impl);
 int mt_dev_if_pre_uinit(struct mtl_main_impl* impl);
 
 int mt_dev_put_lcore(struct mtl_main_impl* impl, unsigned int lcore);
