+++ lib/src/mt_arp.c
@@ -4,7 +4,7 @@
 
 #include "mt_arp.h"
 
-#include "mt_dev.h"
+#include "dev/mt_dev.h"
 // #define DEBUG
 #include "mt_log.h"
 #include "mt_socket.h"
