+++ lib/src/mt_dhcp.c
@@ -4,7 +4,7 @@
 
 #include "mt_dhcp.h"
 
-#include "mt_dev.h"
+#include "dev/mt_dev.h"
 #include "mt_log.h"
 #include "mt_util.h"
