+++ lib/src/mt_ptp.c
@@ -4,8 +4,8 @@
 
 #include "mt_ptp.h"
 
+#include "dev/mt_dev.h"
 #include "mt_cni.h"
-#include "mt_dev.h"
 // #define DEBUG
 #include "mt_log.h"
 #include "mt_mcast.h"
