+++ lib/src/mt_rtcp.c
@@ -5,7 +5,7 @@
 #include "mt_rtcp.h"
 
 #include "datapath/mt_queue.h"
-#include "mt_dev.h"
+#include "dev/mt_dev.h"
 #include "mt_log.h"
 #include "mt_stat.h"
 #include "mt_util.h"
