+++ lib/src/udp/udp_rxq.c
@@ -4,7 +4,7 @@
 
 #include "udp_rxq.h"
 
-#include "../mt_dev.h"
+#include "../dev/mt_dev.h"
 #include "../mt_log.h"
 #include "../mt_stat.h"
 #include "udp_main.h"
